+++ task/Task.js
@@ -1,6 +1,5 @@
 "use strict";
 var Q = require("q");
-var TaskState = require("./TaskState");
 /** Task implementation for a synchronous or asynchronous task
  * @template R the type of data returned by this task if it succeeds
  * @template S the type of error throw by this task if it fails
@@ -11,7 +10,7 @@ var Task = /** @class */ (function () {
     function Task(name, action, dfd) {
         if (dfd === void 0) { dfd = Q.defer(); }
         this.name = name;
-        this.state = TaskState.CREATED;
+        this.state = "CREATED";
         this.action = action;
         this.actionDfd = dfd;
         this.result = undefined;
@@ -19,28 +18,28 @@ var Task = /** @class */ (function () {
     }
     Task.prototype.start = function () {
         var that = this;
-        if (this.state !== TaskState.CREATED) {
+        if (this.state !== "CREATED") {
             throw new Error("task has already been started, cannot start task more than once");
         }
         // resolve/reject run on next tick so 'action' functions that return immediately don't cause TaskSet.getPromises() to return bad results
         function taskCompleted(res) {
-            that.state = TaskState.COMPLETED;
+            that.state = "COMPLETED";
             that.result = res != null ? res : null;
             that.actionDfd.resolve(res);
         }
         function taskErrored(err) {
-            that.state = TaskState.ERRORED;
+            that.state = "ERRORED";
             that.error = err != null ? err : null;
             that.actionDfd.reject(err);
         }
-        this.state = TaskState.AWAITING_EXECUTION;
+        this.state = "AWAITING_EXECUTION";
         if (Task.isPromise(this.action)) {
-            this.state = TaskState.RUNNING; // TODO technically incorrect, we don't know when the task will run in the browser/node/rhino/etc.
+            this.state = "RUNNING"; // TODO technically incorrect, we don't know when the task will run in the browser/node/other.
             this.action.then(taskCompleted, taskErrored);
         }
         else {
             try {
-                this.state = TaskState.RUNNING;
+                this.state = "RUNNING";
                 var res = this.action();
                 taskCompleted(res);
             }
@@ -51,7 +50,7 @@ var Task = /** @class */ (function () {
         return this.actionDfd.promise;
     };
     Task.prototype.isSettled = function () {
-        return this.state.isSettled();
+        return Task.isSettled(this.state);
     };
     Task.prototype.getPromise = function () {
         return this.actionDfd.promise;
@@ -61,6 +60,12 @@ var Task = /** @class */ (function () {
     };
     Task.prototype.getError = function () {
         return this.error;
+    };
+    Task.isSettled = function (state) {
+        return state === "CANCELED" || state === "ERRORED" || state === "COMPLETED";
+    };
+    Task.isRunning = function (state) {
+        return state === "RUNNING" || state === "AWAITING_SCHEDULING" || state === "AWAITING_CHILDREN_COMPLETION" || state === "AWAITING_EXECUTION";
     };
     Task.isPromise = Q.isPromiseAlike;
     return Task;